chore(core): restructure users
