+++ src/bp/core/app.inversify.ts
@@ -4,6 +4,7 @@ import { EventCollector } from 'core/events'
 import { ModuleLoader } from 'core/modules'
 import { TelemetryContainerModules } from 'core/telemetry'
 import { TYPES } from 'core/types'
+import { DataRetentionJanitor, DataRetentionService } from 'core/users'
 import { Container } from 'inversify'
 
 import { BotpressAPIProvider } from './api'
@@ -15,8 +16,6 @@ import { RepositoriesContainerModules } from './repositories/repositories.invers
 import HTTPServer from './server'
 import { LocalActionServer } from './services/action/local-action-server'
 import { MigrationService } from './services/migration'
-import { DataRetentionJanitor } from './services/retention/janitor'
-import { DataRetentionService } from './services/retention/service'
 import { ServicesContainerModules } from './services/services.inversify'
 import { WorkspaceService } from './services/workspace-service'
 import { Statistics } from './stats'
