+++ src/bp/core/botpress.ts
@@ -11,6 +11,7 @@ import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, LogsJanitor } f
 import { copyDir } from 'core/misc/pkg-fs'
 import { ModuleLoader } from 'core/modules'
 import { StatsService } from 'core/telemetry'
+import { DataRetentionJanitor, DataRetentionService } from 'core/users'
 import { WrapErrorsWith } from 'errors'
 import fse from 'fs-extra'
 import { inject, injectable, tagged } from 'inversify'
@@ -44,8 +45,6 @@ import { MigrationService } from './services/migration'
 import { MonitoringService } from './services/monitoring'
 import { NotificationsService } from './services/notification/service'
 import RealtimeService from './services/realtime'
-import { DataRetentionJanitor } from './services/retention/janitor'
-import { DataRetentionService } from './services/retention/service'
 import { WorkspaceService } from './services/workspace-service'
 import { Statistics } from './stats'
 import { TYPES } from './types'
