--- src/bp/core/collaborators/index.ts
@@ -1,3 +1,0 @@
-export * from './repositories/strategy_users'
-export * from './repositories/workspace_users'
-export * from './repositories/workspace_invite_codes'
