+++ src/bp/core/converse/converse-service.ts
@@ -1,8 +1,8 @@
 import { IO } from 'botpress/sdk'
 import { ConfigProvider } from 'core/config/config-loader'
 import { EventEngine } from 'core/events'
-import { UserRepository } from 'core/repositories'
 import { TYPES } from 'core/types'
+import { ChannelUserRepository } from 'core/users'
 import { InvalidParameterError } from 'errors'
 import { EventEmitter2 } from 'eventemitter2'
 import { inject, injectable, postConstruct } from 'inversify'
@@ -32,7 +32,7 @@ export class ConverseService {
   constructor(
     @inject(TYPES.ConfigProvider) private configProvider: ConfigProvider,
     @inject(TYPES.EventEngine) private eventEngine: EventEngine,
-    @inject(TYPES.UserRepository) private userRepository: UserRepository,
+    @inject(TYPES.UserRepository) private userRepository: ChannelUserRepository,
     @inject(TYPES.ConversationService) private conversationService: ConversationService,
     @inject(TYPES.MessageService) private messageService: MessageService
   ) {}
