+++ src/bp/core/database/tables/bot-specific/index.ts
@@ -3,4 +3,3 @@ export * from './notifications'
 export * from './conversations'
 export * from './messages'
 export * from './tasks'
-export * from './users'
