+++ src/bp/core/database/tables/server-wide/index.ts
@@ -1,4 +1,2 @@
-export * from './channel_users'
 export * from './metadata'
-export * from './data_retention'
 export * from './migrations'
