+++ src/bp/core/repositories/index.ts
@@ -1,2 +1,1 @@
 export * from './notifications'
-export * from './users'
