+++ src/bp/core/services/auth/auth-service.ts
@@ -1,7 +1,5 @@
 import { Logger, StrategyUser } from 'botpress/sdk'
 import { JWT_COOKIE_NAME } from 'common/auth'
-import { StrategyUsersRepository } from 'core/collaborators/repositories/strategy_users'
-import { StrategyUserTable } from 'core/collaborators/tables/strategy_users'
 import { AuthStrategy, ConfigProvider } from 'core/config'
 import Database from 'core/database'
 import { SessionIdFactory } from 'core/dialog/sessions'
@@ -11,6 +9,8 @@ import { getMessageSignature } from 'core/misc/security'
 import { ModuleLoader } from 'core/modules'
 import { BadRequestError } from 'core/routers/errors'
 import { Event } from 'core/sdk/impl'
+import { StrategyUsersRepository } from 'core/users'
+import { StrategyUserTable } from 'core/users/tables'
 import { Response } from 'express'
 import { inject, injectable, tagged } from 'inversify'
 import jsonwebtoken from 'jsonwebtoken'
