+++ src/bp/core/services/workspace-service.ts
@@ -10,14 +10,14 @@ import {
 } from 'botpress/sdk'
 import { CHAT_USER_ROLE, defaultPipelines, defaultWorkspace } from 'common/defaults'
 import { AuthRole, CreateWorkspace, Pipeline, Workspace } from 'common/typings'
+import { ConfigProvider } from 'core/config'
+import { ConflictError, NotFoundError } from 'core/routers/errors'
 import {
   StrategyUsersRepository,
   WorkspaceUsersRepository,
   WorkspaceInviteCode,
   WorkspaceInviteCodesRepository
-} from 'core/collaborators'
-import { ConfigProvider } from 'core/config/config-loader'
-import { ConflictError, NotFoundError } from 'core/routers/errors'
+} from 'core/users'
 import { inject, injectable, tagged } from 'inversify'
 import _ from 'lodash'
 import nanoid from 'nanoid/generate'
