+++ src/bp/core/users/chat-users/index.ts
@@ -1,0 +1,1 @@
+export * from './chat-users-repository'
