+++ src/bp/core/users/collaborators/index.ts
@@ -1,0 +1,1 @@
+export * from './repositories'
