+++ src/bp/core/users/tables.ts
@@ -1,0 +1,4 @@
+export * from './chat-users/chat-users-table'
+export * from './chat-users/channel_users-table'
+export * from './collaborators/tables'
+export * from './data-retention/data_retention-table'
